+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - `parse_acl` function to call `parse_acl_item` on each item in a list.
 - Documentation on `parse_acl_item` about ensuring `subname` is a valid
   identifier if it is to be executed.
+- `get_all_column_acls` and `get_column_acls` to get column ACLs.
 
 ### Changed
 - Renamed singular functions like `get_table_acls` to `get_table_acl`.